MED-1269 flash message queue HIGH LOW priority
When the first flash message is added to the messageQueue, it is removed and displayed

Subsequent flash messages are placed on the messageQueue unless the message, status, and priority are the same as the currently displayed flash message

After a delay, the next flash message is dequeued and displayed.

The messageQueue is checked for duplicates of the currently displayed message, and they are removed.

Flash messages with "HIGH" priority are placed at the front of the queue.

Flash messages with "LOW" priority are placed at the tail of the queue. Flash messages with undefined priority are set to LOW priority.