MED-2317_js-code-coverage
Add code coverage for javascript code
Added test to fail if javascript code coverage is below a minimum of 70.0 %, or a minimum set by the **COVERAGE_PERCENT_JS** environ var (must be a number i.e. 77.7)
Had to change some of the keen tests and upgrade simon to get this code to work.
CI will have a report tab with a javascript coverage report.

TODO: I was unable to get karma to work with several code coverage libraries, angular was not initializing correctly. One day we can revisit this